separated styles of the component
