+++ components/MainTextComponent/index.js
@@ -1,0 +1,3 @@
+import MainText from "./MainText";
+
+export default MainText;
