+++ components/MainTextComponent/styles.android.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const stylesAndroid = StyleSheet.create({
+  text: {
+    fontSize: 18,
+    fontFamily: "Roboto",
+    color: "red",
+    textAlign: "center",
+    textDecorationLine: "underline",
+  },
+});
+
+export default stylesAndroid;
