+++ components/MainTextComponent/styles.ios.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const stylesIos = StyleSheet.create({
+  text: {
+    fontSize: 20,
+    fontFamily: "Avenir",
+    color: "red",
+    textAlign: "center",
+    textDecorationLine: "underline",
+  },
+});
+
+export default stylesIos;
