+++ components/MainTextComponent/stylesMainText.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+const stylesMainText = StyleSheet.create({
+  text: {
+    color: "red",
+    textAlign: "center",
+    textDecorationLine: "underline",
+  },
+});
+
+export default stylesMainText;
